moving concrete datasets to cli only
